Refactor OrdersController, add CurrentOrder module, improve order management UI
