+++ app/controllers/orders_controller.rb
@@ -1,7 +1,23 @@
 class OrdersController < ApplicationController
+  include CurrentOrder
+
   before_action :authenticate_user!
-  before_action :set_order, only: [:confirm_order]
-  before_action :authorize_admin, only: [:pending_orders]
+  before_action :set_order, only: [:confirm_order, :update_status]
+  before_action :authorize_admin, only: [:index, :update_status]
+  #
+  def index
+    # The status parameter is expected to come from the URL query string
+    # e.g., /orders?status=pending
+    @status = params[:status]
+
+    @orders = if @status.present?
+                Order.where(status: @status)
+              else
+                Order.all
+              end
+
+    @orders = @orders.order(created_at: :desc)
+  end
 
   def confirm_order
     if @order.created? && @order.update(status: :pending)
@@ -11,15 +27,11 @@ class OrdersController < ApplicationController
     end
   end
 
-  def pending_orders
-    @orders = Order.pending
-  end
-
   def update_status
     if @order.update(order_params)
-      redirect_to pending_orders_orders_path, notice: 'Order status was successfully updated.'
+      redirect_to orders_path, notice: 'Order status was successfully updated.'
     else
-      render :pending_orders
+      render :index
     end
   end
