--- app/views/orders/pending_orders.html.erb
@@ -1,27 +1,0 @@
-<h1>Pending Orders</h1>
-
-<table>
-  <thead>
-    <tr>
-      <th>Order ID</th>
-      <th>User</th>
-      <th>Status</th>
-      <th>Actions</th>
-    </tr>
-  </thead>
-  <tbody>
-    <% @pending_orders.each do |order| %>
-      <tr>
-        <td><%= order.id %></td>
-        <td><%= order.user.email %></td>
-        <td><%= order.status %></td>
-        <td>
-          <%= form_for order, url: update_status_order_path(order), method: :patch do |f| %>
-            <%= f.select :status, Order.statuses.keys %>
-            <%= f.submit "Update Status" %>
-          <% end %>
-        </td>
-      </tr>
-    <% end %>
-  </tbody>
-</table>
