+++ config/routes.rb
@@ -24,6 +24,7 @@ Rails.application.routes.draw do
     end
   end
 
+  get 'orders', to: 'orders#index'
 
   # Define your application routes per the DSL in https://guides.rubyonrails.org/routing.html
